# [nexmark] Add support for ensuring average sizes of people, auctions and bids