Signed-off-by: Michael Nelson <minelson@vmware.com>

This PR adds the missing functionality of ensuring that each event (person, auction or bid) matches (on average) the configured average size in bytes for that record.

As expected (extra strings), it hits the performance slightly, adding around and extra second to the processing of 10M events for q4.

## Before

```
cargo bench --bench nexmark --features with-nexmark -- --first-event-rate=10000000 --max-events=10000000 --cpu-cores 4 --query q4 --num-event-generators 4 --source-buffer-size 10000 --input-batch-size 10000
   Compiling dbsp v0.1.0 (/home/michael/dev/vmware/database-stream-processor)
    Finished bench [optimized + debuginfo] target(s) in 35.22s
     Running benches/nexmark/main.rs (target/release/deps/nexmark-c782d17920c2b5ef)
Starting q4 bench of 10000000 events...
Done                                                                                                                                                                                                                ┌───────┬────────────┬───────┬─────────┬─────────────────┬──────────────────┬───────────────┬───────────────┬─────────────┐
│ Query │ #Events    │ Cores │ Elapsed │ Cores * Elapsed │ Throughput/Cores │ Total Usr CPU │ Total Sys CPU │ Max RSS(Kb) │
├───────┼────────────┼───────┼─────────┼─────────────────┼──────────────────┼───────────────┼───────────────┼─────────────┤
│ q4    │ 10,000,000 │ 4     │ 3.391s  │ 13.565s         │ 737.203 K/s      │ 16.935s       │ 0.462s        │ 566,668     │
└───────┴────────────┴───────┴─────────┴─────────────────┴──────────────────┴───────────────┴───────────────┴─────────────┘
```

## After
```
$ cargo bench --bench nexmark --features with-nexmark -- --first-event-rate=10000000 --max-events=10000000 --cpu-cores 4 --query q4 --num-event-generators 4 --source-buffer-size 10000 --input-batch-size 10000
    Finished bench [optimized + debuginfo] target(s) in 0.10s
     Running benches/nexmark/main.rs (target/release/deps/nexmark-c782d17920c2b5ef)
Starting q4 bench of 10000000 events...
Done                                                                                                                                                                                                                ┌───────┬────────────┬───────┬─────────┬─────────────────┬──────────────────┬───────────────┬───────────────┬─────────────┐
│ Query │ #Events    │ Cores │ Elapsed │ Cores * Elapsed │ Throughput/Cores │ Total Usr CPU │ Total Sys CPU │ Max RSS(Kb) │
├───────┼────────────┼───────┼─────────┼─────────────────┼──────────────────┼───────────────┼───────────────┼─────────────┤
│ q4    │ 10,000,000 │ 4     │ 4.448s  │ 17.793s         │ 562.031 K/s      │ 25.061s       │ 0.975s        │ 608,756     │
└───────┴────────────┴───────┴─────────┴─────────────────┴──────────────────┴───────────────┴───────────────┴─────────────┘
```

Actually, you can just run the test so that no extra string is generated by setting the `--avg-*-byte-size` options to zero:

```
cargo bench --bench nexmark --features with-nexmark -- --first-event-rate=10000000 --max-events=10000000 --cpu-cores 4 --query q4 --num-event-generators 6 --source-buffer-size 10000 --input-batch-size 10000 --avg-person-byte-size 0 --avg-auction-byte-size 0 --avg-bid-byte-size 0
    Finished bench [optimized + debuginfo] target(s) in 0.11s
     Running benches/nexmark/main.rs (target/release/deps/nexmark-c782d17920c2b5ef)
Starting q4 bench of 10000000 events...
Done                                                                                                                                                                                                                ┌───────┬────────────┬───────┬─────────┬─────────────────┬──────────────────┬───────────────┬───────────────┬─────────────┐
│ Query │ #Events    │ Cores │ Elapsed │ Cores * Elapsed │ Throughput/Cores │ Total Usr CPU │ Total Sys CPU │ Max RSS(Kb) │
├───────┼────────────┼───────┼─────────┼─────────────────┼──────────────────┼───────────────┼───────────────┼─────────────┤
│ q4    │ 10,000,000 │ 4     │ 3.050s  │ 12.200s         │ 819.641 K/s      │ 14.447s       │ 0.571s        │ 608,988     │
└───────┴────────────┴───────┴─────────┴─────────────────┴──────────────────┴───────────────┴───────────────┴─────────────┘
```